+++ src/cli/commands/install.rs
@@ -59,9 +59,9 @@ pub struct InstallArgs {
     #[arg(long, default_value = "false")]
     skip_test: bool,
 
-    /// True to skip the build tests (note that these are not the same as the Packit tests)
+    /// True to execute the build tests (note that these are not the same as the Packit tests)
     #[arg(long, default_value = "false")]
-    include_build_test: bool,
+    execute_build_test: bool,
 
     /// True to pause the install after the build is completed (to debug builds)
     #[arg(long, default_value = "false")]
@@ -120,7 +120,7 @@ impl HandleCommand for InstallArgs {
             .skip_active(self.skip_active)
             .keep_build(self.keep_build)
             .verbose(self.verbose)
-            .include_build_test(self.include_build_test)
+            .execute_build_test(self.execute_build_test)
             .skip_test(self.skip_test)
             .pause_build(self.pause_build);
         let mut installer = Installer::new(&config, &mut register, &manager, installer_options);