+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - New and improved documentation.
 
 ### Changes
-- The build tests are now turned off by default and can be turned on with `--include-build-test` (`--skip-build-test` is removed).
+- The build tests are now turned off by default and can be turned on with `--execute-build-test` (`--skip-build-test` is removed).
 
 ### Fixed
 - Fix patch apply directory meta check reporting wrong issues.